backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
passlib
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
bcrypt
pydantic[email]
motor
pymongo
boto3